Add TODO comment and update mock data in unit tests
Added a TODO comment in the MedicalRecordService to handle a case where birthdate could be in the future in the createRecord method. Also updated mock test data in ManageMockedData for the unit test, updating the value of the medical record from Maxim to mini-Maxim and changing the birthdate to enhance test coverage.