Introduced request logging
A new RequestLoggingFilterConfig class has been added to log every inbound HTTP request. This request logging will include query string, payload up to 1000 characters, but not headers. The log level for the new CommonsRequestLoggingFilter class is set to DEBUG in 'application.properties'.

The logging enhancement is introduced to provide debugging information and the addition to the 'persons.json' is to bring more data into the application.