Implement logic to load data from dataset.json on server initiation
A new service, InitialLoadDataService, is introduced to read an initial dataset from an .json file. This service reads the dataset only once at the start of the server. Adding this functionality improves the efficiency of data loading and segregating data related to Persons, Firestations, and MedicalRecords into different repositories. Unit tests for the service have also been added to verify its successful operation. This commit also includes the creation of a new DTO, LoadInitialDataDTO, to transfer data from the service to the repositories.