Add data initialization in tests
Added a command line runner to initialize data for dev and devpartial profiles.
Made changes to use initialLoadDataService for initializing and clearing data in all ITs.
Updated savePerson method in PersonRepository to avoid concurrent modification Exception.
Refactored the test configuration for utilizing the partial and full mock datasets.

This was required to be able to test with realistic data and avoid hardcoding values in the tests.
Also, initializing and clearing data in a uniform way makes the test code cleaner.